[REF] joint_buying_base : split test into many files
