+++ joint_buying_base/tests/__init__.py
@@ -1,3 +1,7 @@
-from . import test_module
+from . import test_abstract
 from . import test_company_2_partner
+from . import test_joint_buying_check_access_mixin
+from . import test_joint_buying_mixin
+from . import test_partner_global_local
+from . import test_partner_subscription
 from . import test_tour
