+++ joint_buying_base/tests/test_partner_global_local.py
@@ -1,0 +1,39 @@
+# Copyright (C) 2021 - Today: GRAP (http://www.grap.coop)
+# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from odoo.exceptions import ValidationError
+from odoo.tests import tagged
+
+from .test_abstract import TestAbstract
+
+
+@tagged("post_install", "-at_install")
+class TestModule(TestAbstract):
+    def setUp(self):
+        super().setUp()
+        self.company_ELD = self.env.ref("joint_buying_base.company_ELD")
+        self.company_CHE = self.env.ref("joint_buying_base.company_CHE")
+        self.company_3PP = self.env.ref("joint_buying_base.company_3PP")
+
+    # Test Section
+    def test_601_double_link_supplier_to_joint_buying_partner(self):
+        # Create a new supplier in a company linked to a joint buying partner should
+        # success
+        partner_id = self.company_ELD.joint_buying_partner_id.id
+        vals = {
+            "name": "Test Elodie-D @ CHE",
+            "company_id": self.company_CHE.id,
+            "supplier": True,
+            "joint_buying_global_partner_id": partner_id,
+        }
+        self.ResPartner.create(vals)
+
+        with self.assertRaises(ValidationError):
+
+            # We should not have the possibility to link two suppliers
+            # to the same joint buying supplier for the same company
+            vals.update(
+                {"name": "Test Elodie-D @ 3PP", "company_id": self.company_3PP.id}
+            )
+            self.ResPartner.create(vals)
